model.CreateResource does not check for the existence of blank ids.
SPARQL 1.1 spec 5.1.1: When using blank nodes of the form _:abc,  labels for blank nodes are scoped to the basic graph pattern.

This renders checking for the existence of blank node ids upon creation of resources highly unreliable. Therefore skip the check in this case as the blank id will be reassigned to the actual (store provided) values after Commit() was invoked.