Changed model blank node retrieval behavior and minor bugfix.
Changed model behavior: Resources that are identified using blank node ids cannot be retrieved from the store using the URI. This is because SPARQL 1.1 officially treats blank nodes as variables and hence does not support using blank nodes as subjects in FILTERS, VALUEs or DESCRIBE queries.

Bugfix in dotNetRDF store: When retrieving objects that reference resources that are identified by a blank node, then the referenced resources URIs are now properly restored as blank node id.